LayoutTests/imported/w3c/web-platform-tests/shadow-dom/reference-target/tentative/reference-target-basics-expected.txt


FAIL ShadowRoot.referenceTarget defaults to null when shadow is created declaratively assert_equals: expected (object) null but got (string) ""
PASS Empty shadowrootreferencetarget attribute is reflected as empty string
PASS <template> shadowrootreferencetarget sets referenceTarget on shadow root
FAIL ShadowRoot.referenceTarget defaults to null when shadow is created imperatively assert_equals: expected (object) null but got (string) ""
PASS Passing empty string referencetarget in ShadowRootInit is reflected as empty string
PASS ShadowRootInitDict can be used to set referenceTarget on shadow root
